fixed dropping tables problem in elephantsql

+++ server.py
@@ -759,10 +759,12 @@ def create_tables():
 def drop_tables():
     with dbapi2.connect(app.config['dsn']) as connection:
                 cursor = connection.cursor()
-                cursor.execute("SELECT table_schema,table_name FROM information_schema.tables WHERE table_schema = 'public' ORDER BY table_schema,table_name")
+                cursor.execute("""SELECT table_schema,table_name FROM information_schema.tables 
+                                WHERE table_schema = 'public' ORDER BY table_schema,table_name""")
                 rows = cursor.fetchall()
                 for row in rows:
-                    cursor.execute("drop table " + row[1] + " cascade")
+                    if row[1]!="pg_stat_statements":
+                        cursor.execute("drop table " + row[1] + " cascade")
 
                 connection.commit()
     return redirect(url_for('create_tables'))
